Moved bundle components into a subfolder.
