+++ src/components/bundles/BundleItems.vue
